python-trajectory/projects/Multi Agent chat  /qt-multiagent-chat/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "qt-multiagent-chat"
version = "0.1.0"
description = "Qt-based multi-agent chat interface with Ollama integration"
dependencies = [
    "PySide6>=6.5.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
qt_api = "pyside6"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
